--Saco la información del problema a una clase Problema. --Cambio el nombre de la clase ProblemaMochila por EjecutorProblemaMochila --Leo la instancia del problema por fichero.

+++ src/es/uned/mochila/Objeto.java
@@ -13,7 +13,7 @@ public class Objeto {
 		this.volumen=Math.random() * (maxValor-minValor) + minValor;
 	}
 	
-	public Objeto(float valor, float volumen){
+	public Objeto(double valor, double volumen){
 		this.valor=valor;
 		this.volumen=volumen;
 	}
